use a wrapper to the InputStream as payload of SimpleStorage (used by the tests) so the input-stream is only created when it is really used. avoids hanging input-streams during the tests.
moved the layout tests into the layout package
