+++ plugins/rubygem/nexus-ruby-tools/src/test/java/org/sonatype/nexus/ruby/layout/DELETELayoutTest.java
@@ -10,10 +10,13 @@
  * of Sonatype, Inc. Apache Maven is a trademark of the Apache Software Foundation. M2eclipse is a trademark of the
  * Eclipse Foundation. All other trademarks are the property of their respective owners.
  */
-package org.sonatype.nexus.ruby;
+package org.sonatype.nexus.ruby.layout;
 
+import java.io.File;
+
+import org.sonatype.nexus.ruby.FileType;
+import org.sonatype.nexus.ruby.RubygemsFile;
 import org.sonatype.nexus.ruby.cuba.DefaultRubygemsFileSystem;
-import org.sonatype.nexus.ruby.layout.NoopDefaultLayout;
 import org.sonatype.sisu.litmus.testsupport.TestSupport;
 
 import org.junit.Test;
@@ -22,11 +25,12 @@ import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.is;
 
-public class NoopDefaultLayoutTest
+public class DELETELayoutTest
     extends TestSupport
 {
-  private final DefaultRubygemsFileSystem bootstrap = new DefaultRubygemsFileSystem(
-      new NoopDefaultLayout(null, null), null, null);
+  private final DefaultRubygemsFileSystem bootstrap =
+      new DefaultRubygemsFileSystem(new DELETELayout(null, new SimpleStorage(new File("target"))),
+          null, null);
 
   @Test
   public void testSpecsZippedIndex() throws Exception {
@@ -35,17 +39,17 @@ public class NoopDefaultLayoutTest
         "/prerelease_specs.4.8.gz",
         "/latest_specs.4.8.gz"
     };
-    assertFiletype(pathes, FileType.SPECS_INDEX_ZIPPED);
+    assertFound(pathes, FileType.SPECS_INDEX_ZIPPED);
   }
 
   @Test
-  public void testSpecsIndex() throws Exception {
+  public void testSpecsUnzippedIndex() throws Exception {
     String[] pathes = {
         "/specs.4.8",
         "/prerelease_specs.4.8",
         "/latest_specs.4.8"
     };
-    assertFiletype(pathes, FileType.SPECS_INDEX);
+    assertForbidden(pathes);
   }
 
   @Test
@@ -104,7 +108,7 @@ public class NoopDefaultLayoutTest
   @Test
   public void testApiV1() throws Exception {
     String[] pathes = {"/api/v1/gems", "/api/v1/api_key"};
-    assertFiletype(pathes, FileType.API_V1);
+    assertForbidden(pathes);
   }
 
   @Test
@@ -113,26 +117,26 @@ public class NoopDefaultLayoutTest
         "/api/v1/dependencies?gems=nexus", "/api/v1/dependencies/jbundler.json.rz",
         "/api/v1/dependencies/b/bundler.json.rz"
     };
-    assertFiletype(pathes, FileType.DEPENDENCY);
+    assertFound(pathes, FileType.DEPENDENCY);
   }
 
   @Test
   public void testGemspec() throws Exception {
     String[] pathes = {"/quick/Marshal.4.8/jbundler.gemspec.rz", "/quick/Marshal.4.8/b/bundler.gemspec.rz"};
-    assertFiletype(pathes, FileType.GEMSPEC);
+    assertFound(pathes, FileType.GEMSPEC);
   }
 
   @Test
   public void testGem() throws Exception {
     String[] pathes = {"/gems/jbundler.gem", "/gems/b/bundler.gem"};
-    assertFiletype(pathes, FileType.GEM);
+    assertFound(pathes, FileType.GEM);
   }
 
   @Test
   public void testDirectory() throws Exception {
     String[] pathes = {
         "/", "/api", "/api/", "/api/v1", "/api/v1/",
-        "/api/v1/dependencies", "/gems/", "/gems"
+        "/api/v1/dependencies", "/gems/", "/gems",
     };
     assertForbidden(pathes);
     String[] mpathes = {
@@ -172,7 +176,18 @@ public class NoopDefaultLayoutTest
 
   protected void assertFiletype(String[] pathes, FileType type) {
     for (String path : pathes) {
-      assertThat(path, bootstrap.get(path).type(), equalTo(type));
+      RubygemsFile file = bootstrap.get(path);
+      assertThat(path, file.type(), equalTo(type));
+      assertThat(path, file.get(), equalTo(null));
+      assertThat(path, file.hasException(), is(false));
+    }
+  }
+
+  protected void assertFound(String[] pathes, FileType type) {
+    for (String path : pathes) {
+      RubygemsFile file = bootstrap.get(path);
+      assertThat(path, file.type(), equalTo(type));
+      assertThat(path, file.notExists(), is(false));
     }
   }
